third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-ui/show-function-definition-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that "Show Function Definition" jumps to the correct location.


Running: testRevealFunctionDefinition
Function location revealed: [10:24]

Running: testDumpFunctionDefinition
jumpToMe
ƒ jumpToMe()
      {
          var result = 12345;
          return window.foo || result;
      }
